refactor to allow centroid distance sorting to be optional
