+++ query/geo_base.js
@@ -1,12 +1,13 @@
 
 // Base Query for GeoCoding Queries
 
-module.exports = function( opts ){
+module.exports = function( centroid, opts ){
 
   if( !opts ){ opts = {}; }
 
   var options = {
-    size: opts.size || 1
+    size: opts.size || 1,
+    sort: opts.sort || false
   };
   
   var query = {
@@ -28,6 +29,19 @@ module.exports = function( opts ){
     ],
     "track_scores": true
   };
+
+  // sort results by distance from input centroid
+  if( options.sort && centroid ){
+    var sort = [{
+      '_geo_distance': {
+        'center_point': centroid,
+        'order': 'asc',
+        'unit': 'km'
+      }
+    }];
+
+    query.sort = query.sort.concat(sort);
+  }
 
   return query;
 }
